requirements.txt
python-dotenv
mutagen
spotipy
requests
